+++ src/chains/supportedChains.evm.int.spec.ts
@@ -1,7 +1,6 @@
 import { createPublicClient, http } from 'viem'
 import { describe, expect, test } from 'vitest'
-import { supportedEVMChains } from './supportedChains.evm'
-import { isSameUrl } from './utils'
+import { supportedEVMChains } from './supportedChains.evm.js'
 
 describe.concurrent('EVM chains RPC check', () => {
   const rpcUrls = supportedEVMChains.flatMap((chain) =>
@@ -22,28 +21,6 @@ describe.concurrent('EVM chains RPC check', () => {
 
       const _chainId = await client.getChainId()
       expect(_chainId).toBe(chainId)
-    }
-  )
-})
-
-describe.concurrent('EVM chains block explorer check', () => {
-  const blockExplorerUrls = supportedEVMChains.flatMap((chain) =>
-    chain.metamask.blockExplorerUrls.map((blockExplorerUrl) => ({
-      blockExplorerUrl: blockExplorerUrl,
-      chainId: chain.id,
-      chainName: chain.name,
-    }))
-  )
-
-  test.for(blockExplorerUrls)(
-    `block explorer should be alive $chainName - $chainId - $blockExplorerUrl`,
-    { timeout: 10_000, retry: 3 },
-    async ({ blockExplorerUrl }) => {
-      const response = await fetch(blockExplorerUrl)
-
-      expect(isSameUrl(blockExplorerUrl, response.url)).toBeTruthy()
-      expect(response.ok).toBe(true)
-      expect(response.status).toBe(200)
     }
   )
 })