+++ src/chains/supportedChains.svm.int.spec.ts
@@ -1,7 +1,6 @@
 import { Connection, PublicKey } from '@solana/web3.js'
 import { describe, expect, test } from 'vitest'
-import { supportedSolanaChains } from './supportedChains.svm'
-import { isSameUrl } from './utils'
+import { supportedSolanaChains } from './supportedChains.svm.js'
 
 const TokenProgramAddress = 'TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA'
 const WalletAddress = '6AUWsSCRFSCbrHKH9s84wfzJXtD6mNzAHs11x6pGEcmJ'
@@ -46,27 +45,6 @@ describe.concurrent('SVM chains RPC check', () => {
       expect(slot.status).toBe('fulfilled')
       expect(balance.status).toBe('fulfilled')
       expect(tokenAccountsByOwner.status).toBe('fulfilled')
-    }
-  )
-})
-
-describe.concurrent('SVM chains block explorer check', () => {
-  const blockExplorerUrls = supportedSolanaChains.flatMap((chain) =>
-    chain.metamask.blockExplorerUrls.map((blockExplorerUrl) => ({
-      blockExplorerUrl: blockExplorerUrl,
-      chainId: chain.id,
-      chainName: chain.name,
-    }))
-  )
-
-  test.for(blockExplorerUrls)(
-    `block explorer should be alive $chainName - $chainId - $blockExplorerUrl`,
-    { timeout: 10_000, retry: 3 },
-    async ({ blockExplorerUrl }) => {
-      const response = await fetch(blockExplorerUrl)
-      expect(isSameUrl(blockExplorerUrl, response.url)).toBeTruthy()
-      expect(response.ok).toBe(true)
-      expect(response.status).toBe(200)
     }
   )
 })