+++ src/coins/coins.int.spec.ts
@@ -1,6 +1,6 @@
 import type { StaticToken } from '@lifi/types'
 import { describe, expect, test } from 'vitest'
-import { defaultCoins, wrappedTokens } from './coins'
+import { defaultCoins, wrappedTokens } from './coins.js'
 
 describe.concurrent('Coin logo test', { timeout: 30_000 }, () => {
   const allImages: string[] = []
@@ -23,7 +23,7 @@ describe.concurrent('Coin logo test', { timeout: 30_000 }, () => {
     ...Object.values(wrappedTokens).map((token: StaticToken) => token.logoURI!)
   )
 
-  test.each(Array.from(new Set(allImages)))(
+  test.each([...new Set(allImages)])(
     'check that logoURI %s is valid',
     async (image) => {
       expect(image.startsWith('https://')).toBeTruthy()