+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './chains'
-export * from './coins'
-export * from './multicall'
+export * from './chains/index.js'
+export * from './coins/index.js'
+export * from './multicall.js'